refactor(#7) - 피드백 반영
- 테스트 코드 수정
- 낙관적 락 -> 비관적 락 변경
- 불필요한 코드 삭제
