+++ src/main/java/com/flab/infrun/coupon/domain/CouponRepository.java
@@ -9,7 +9,5 @@ public interface CouponRepository {
 
     List<Coupon> saveAll(final List<Coupon> coupons);
 
-    Optional<Coupon> findByCouponCode(final String couponCode);
-
     Optional<Coupon> findByCouponCodeWithLock(final String couponCode);
 }
