+++ src/main/java/com/flab/infrun/coupon/infrastructure/persistence/CouponRepositoryAdapter.java
@@ -22,11 +22,6 @@ public class CouponRepositoryAdapter implements CouponRepository {
     }
 
     @Override
-    public Optional<Coupon> findByCouponCode(final String couponCode) {
-        return couponJpaRepository.findByCode(couponCode);
-    }
-
-    @Override
     public Optional<Coupon> findByCouponCodeWithLock(final String couponCode) {
         return couponJpaRepository.findByCodeWithLock(couponCode);
     }
