+++ src/main/java/com/flab/infrun/coupon/presentation/CouponController.java
@@ -5,6 +5,7 @@ import com.flab.infrun.common.config.security.UserAdapter;
 import com.flab.infrun.common.response.Response;
 import com.flab.infrun.coupon.application.CouponFacade;
 import com.flab.infrun.coupon.presentation.request.CreateCouponRequest;
+import com.flab.infrun.coupon.presentation.request.EnrollCouponRequest;
 import com.flab.infrun.coupon.presentation.response.CreatedCouponResponse;
 import com.flab.infrun.coupon.presentation.response.EnrolledCouponResponse;
 import jakarta.validation.Valid;
@@ -12,7 +13,6 @@ import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -46,11 +46,5 @@ public class CouponController {
         var result = facade.enrollCoupon(request.toCommand(user), LocalDateTime.now());
 
         return Response.success(EnrolledCouponResponse.from(result));
-    }
-
-    @GetMapping
-    @PreAuthorize("hasRole('TEACHER')")
-    public String test() {
-        return "TEST";
     }
 }
